+++ pull_request_template.md
@@ -1,3 +1,4 @@
+- [ ] check Proptypes for edited/new components are properly defined
 - [ ] npm run prettier
 - [ ] npm run storybook : no warning or errors on browser console
 - [ ] check that cooking-with-amateurs works with new build without warning or errors on browser console
