+++ src/components/molecules/recipe/list/toolbar.stories.jsx
@@ -7,10 +7,6 @@ import theme from '../../../../styles/global-styles';
 
 import allLiterals from '../../../../../.storybook/fake_data/literals.json';
 import themeDecorator from '../../../../../.storybook/decorators/themeDecorator';
-import authorData from '../../../../../.storybook/fake_data/author.json';
-import projectData from '../../../../../.storybook/fake_data/project.json';
-
-import escapingBoredomTitleLogo from '../../../../../public/escaping-boredom-title-logo.png';
 
 import Component from './toolbar';
 
@@ -18,9 +14,6 @@ storiesOf('Molecules/[Recipe]/List/Toolbar', module)
   .addDecorator(themeDecorator(theme))
   .add('default', () => (
     <Component
-      authorData={authorData}
-      projectData={projectData}
-      logo={escapingBoredomTitleLogo}
       searchValue="Search Value"
       handleChange={action('Input detected')}
       languageData={{
