+++ src/components/templates/index.js
@@ -1,5 +1,6 @@
 export { default as ItemListTemplate } from './item/list.jsx';
-export { default as RecipeListTemplate } from './recipe/list.jsx';
 export { default as ItemDetailTemplate } from './item/detail.jsx';
+export { default as RecipeListTemplate } from './recipe/list.jsx';
 export { default as RecipeDetailTemplate } from './recipe/detail.jsx';
+export { default as RecipeAboutTemplate } from './recipe/about.jsx';
 export { default as NotFoundTemplate } from './notFound.jsx';
