+++ src/components/templates/recipe/list.stories.jsx
@@ -5,7 +5,6 @@ import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
 
 import literals from '../../../../.storybook/fake_data/literals.json';
-import project from '../../../../.storybook/fake_data/project.json';
 import recipes from '../../../../.storybook/fake_data/recipes.json';
 
 import recipeImage from '../../../../public/fake_imgs/recipe.jpeg';
@@ -26,16 +25,6 @@ storiesOf('Templates/[Recipe]/List', module)
   .add('default', () => (
     <Component
       literals={getLiterals(literals)}
-      authorData={{
-        name: literals.gabriel,
-        email: 'romay.gabriel@gmail.com',
-        url: literals.cv,
-      }}
-      projectData={{
-        description: project.description,
-        url: 'https://github.com/W01fw00d/chemistry-ui/blob/master/README.md',
-        participants: project.participants,
-      }}
       search={literals.comingSoon}
       itemList={formattedRecipes}
       languageData={{
